+++ lib/galaxy/tools/verify/asserts/text.py
